Change in ci/cd and fixes in CM
